requirements.txt
click==8.3.0
dnspython==2.8.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
-e git+ssh://git@github.com/aetheric-forge/parallel-you.git@8355c7ea7ca6622fd362a40afeec71b38e07f863#egg=parallel_you&subdirectory=../../../parallel-you.worktrees/v0.2.0
platformdirs==4.5.0
Pygments==2.19.2
pymongo==4.15.3
PyYAML==6.0.3
rich==14.2.0
shellingham==1.5.4
textual==6.5.0
typer==0.20.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
